list, update, delete for admin paintings

+++ src/main/java/org/viktoriianikitiuk/artoffreedom/dao/PaintingRepository.java
@@ -3,5 +3,5 @@ package org.viktoriianikitiuk.artoffreedom.dao;
 import org.springframework.data.repository.CrudRepository;
 import org.viktoriianikitiuk.artoffreedom.Painting;
 
-public interface PaintingRepository extends CrudRepository<Painting, Integer>{
+public interface PaintingRepository extends CrudRepository<Painting, Long>{
 }
